backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
motor==3.4.0
pymongo==4.7.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
groq==0.11.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

